+++ CHANGELOG.md
@@ -49,7 +49,7 @@
 * Adding `xan parallel --run` & `xan cat rows --run`.
 * Adding `xan to md -l/--limit`.
 * Adding the `xan spark` command.
-* Adding `xan stats -D/--describe, --color, --cols, --sep`.
+* Adding `xan stats -R/--report, --color, --cols, --sep`.
 * Adding `xan (freq|p freq) -X/--approx-algo`.
 
 *Fixes*